Session-3/requirements.txt
trimesh
Pillow
numpy
scipy
nltk
wordnet
torchvision
torch
scipy
scipy.io
scipy.signal    
matplotlib
mpl_toolkits
numpy
